Cover madness
So here's the deal: those pesky buttons are working again, and the file drop appears to be working its magic pretty well. However, for some reason, the coversave.php keeps throwing a 500 internal server error. The files are indeed uploading, and because the application type is declared, I don't think that it's something to do with that. Maybe in the request? I'm sure you'll figure it out, have a wonderful time at Kung Fu Burger!